MatchWarn, Naming, Bugs, Config hierarchies
Added MatchWarn, same as MatchError but different log output
MatchWarn is intended for just warnings to users for whatever might be happening. E.g colliding your own ships in halite, scoring an own goal. Basically, allowed behavior that might not be intended

MatchError is more for actual errors that make break an agent or game

Allow to name tournaments
Allow to apply configs to tourneys better